Refactored, created token service
