+++ core/middlewares/withCurrentUser.js
@@ -1,7 +1,7 @@
 import { STATUS_CODES } from 'http';
 
 import { SESSION } from '@core/constants';
-import { getUserById, verifyToken } from '@core/services/User';
+import { verifyToken } from '@core/services/Auth';
 
 import withToken from './withToken';
 
@@ -9,12 +9,13 @@ export default [
   withToken,
   async (req, res, next) => {
     const { token } = req;
-    const verifiedToken = await verifyToken(token);
-    if (!verifiedToken) {
+    try {
+      const verifiedToken = await verifyToken(token);
+      req.user = verifiedToken;
+      next();
+    } catch (err) {
       res.clearCookie(SESSION.COOKIE_NAME, { signed: true, httpOnly: true });
       return res.status(401).json({ error: STATUS_CODES[401] });
     }
-    req.user = verifiedToken;
-    next();
   },
 ];
